# contracts/openapi/smartmatch.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "description": "The inner error object.",
        "properties": {
          "code": {
            "description": "Stable machine-readable code; safe to branch on",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured context",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable summary; not for branching",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "The response body for every non-2xx API response.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobState": {
        "description": "Every state a durable job may occupy.",
        "enum": [
          "queued",
          "dispatched",
          "running",
          "succeeded",
          "partial",
          "failed_provider",
          "failed_budget",
          "failed_policy",
          "cancelled",
          "timed_out",
          "redrive_pending",
          "abandoned"
        ],
        "title": "JobState",
        "type": "string"
      },
      "JobStatusResponse": {
        "description": "A job's current state.",
        "properties": {
          "command_type": {
            "title": "Command Type",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latest_sequence": {
            "default": 0,
            "description": "Highest job_event sequence emitted so far",
            "title": "Latest Sequence",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/JobState"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "command_type",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "JobStatusResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "IA West SmartMatch platform API \u2014 Foundation scaffold. OpenAPI is the source of truth; the TypeScript client is generated from it and never hand-maintained.",
    "title": "SmartMatch API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/jobs/{job_id}": {
      "get": {
        "description": "Return one job's current state.\n\nTenant-scoped by the caller's own tenant, which is derived server-side from\ntheir token and never read from the request.",
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Get a job's status",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/events": {
      "get": {
        "description": "Stream a job's events as Server-Sent Events.\n\nReconnect is the interesting case. A browser that loses the connection\nreissues the request with ``Last-Event-ID`` set to the last sequence it saw,\nand the stream resumes from there \u2014 no duplicates, no gap. Because the cursor\nis a database column rather than in-memory state, the reconnect can land on\nany API instance.\n\n``after`` exists for clients that cannot set headers (EventSource\npolyfills, simple polling loops) and means the same thing. When both are\npresent ``Last-Event-ID`` wins, since it is the standard.\n\nBrowser disconnect does not cancel the underlying work. Cancellation is an\nexplicit, persisted, authorized command (v1.1 \u00a71.6) \u2014 closing a tab must\nnever stop a running job.",
        "operationId": "stream_job_events_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "description": "Resume cursor for polling clients that cannot send headers",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Resume cursor for polling clients that cannot send headers",
              "title": "After"
            }
          },
          {
            "description": "Resume from this sequence. Standard SSE reconnect header.",
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Resume from this sequence. Standard SSE reconnect header.",
              "title": "Last-Event-Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Stream a job's events (SSE)",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}